Main AI bits and script for interaction texts uploaded. Will be adding more once have art.
The main AI pieces have been summarized into 3 scripts, one for each enemy type, I imagine these will be expanding quite a lot. There is also the interactions text script, which once added to an entity will hold the 6 different phrases for said entity. Once we have the art, I'll be able to add the scripts correctly and expand to cover any animations that are needed for the enemies.